% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlleles.R
\name{addAlleles}
\alias{addAlleles}
\title{Add Alleles}
\usage{
addAlleles(data, ind, line, divider)
}
\arguments{
\item{data}{data}

\item{ind}{ind}

\item{line}{line}

\item{divider}{divider}
}
\value{
data (after alleles were added)
}
\description{
Add Alleles
}
