% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laskeMuutokset12345.R
\name{admix1_muutokset}
\alias{admix1_muutokset}
\title{Calculate changes (?)}
\description{
Palauttaa npops*npops taulun, jonka alkio (i,j) kertoo, mik?on
muutos logml:ss? mikäli populaatiosta i siirretään osuuden verran
todennäköisyysmassaa populaatioon j. Mikäli populaatiossa i ei ole mitään
siirrettävää, on vastaavassa kohdassa rivi nollia.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-admix1_muutokset-laskeMuutokset4}{\code{admix1_muutokset$laskeMuutokset4()}}
\item \href{#method-admix1_muutokset-clone}{\code{admix1_muutokset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-admix1_muutokset-laskeMuutokset4"></a>}}
\if{latex}{\out{\hypertarget{method-admix1_muutokset-laskeMuutokset4}{}}}
\subsection{Method \code{laskeMuutokset4()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{admix1_muutokset$laskeMuutokset4(osuus, osuusTaulu, omaFreqs, logml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{osuus}}{Percentages?}

\item{\code{osuusTaulu}}{Percentage table?}

\item{\code{omaFreqs}}{own Freqs?}

\item{\code{logml}}{log maximum likelihood}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-admix1_muutokset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-admix1_muutokset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{admix1_muutokset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
