% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture_initialization.R
\name{admixture_initialization}
\alias{admixture_initialization}
\title{Seuraavat kolme funktiota liittyvat alkupartition muodostamiseen.}
\usage{
admixture_initialization(data_matrix, nclusters, Z)
}
\description{
Seuraavat kolme funktiota liittyvat alkupartition muodostamiseen.
}
