% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanks.R
\name{blanks}
\alias{blanks}
\title{Blanks}
\usage{
blanks(n)
}
\arguments{
\item{n}{length of vector}
}
\value{
Vector of n blanks
}
\description{
Create character vector of blanks
}
\details{
This function emulates the behavior of a homonimous function from Matlab
}
\author{
Waldir Leoncio
}
