% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopLogml.R
\name{calculatePopLogml}
\alias{calculatePopLogml}
\title{Calculate log marginal likelihood}
\usage{
calculatePopLogml(points, fii)
}
\arguments{
\item{points}{points}

\item{fii}{fii}
}
\description{
Calculates fuzzy (log) marginal likelihood for a population of 
real values using estimate "fii" for the dispersion value, and Jeffreys prior
for the mean parameter.
}
