% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{cell}
\alias{cell}
\title{Cell array}
\usage{
cell(n, sz = c(n, n), expandable = FALSE, ...)
}
\arguments{
\item{n}{a the first dimension (or both, if sz is not passed)}

\item{sz}{the second dimension (or 1st and 2nd, if not passed)}

\item{expandable}{if TRUE, output is a list (so it can take different
lengths)}

\item{...}{Other dimensions}
}
\value{
An array of zeroes with the dimensions passed on call
}
\description{
Creates an array of zeros
}
