% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon.R
\name{colon}
\alias{colon}
\title{Vector creation}
\usage{
colon(a, b)
}
\arguments{
\item{a}{initial number}

\item{b}{final number}
}
\description{
Simulates the function `colon()` and its equivalent `:` operator from Matlab, which have a similar but not quite equivalent behavior when compared to `seq()` and `:` in R.
}
