% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAllFreqs2.R
\name{computeAllFreqs2}
\alias{computeAllFreqs2}
\title{Compute all freqs - version 2}
\usage{
computeAllFreqs2(noalle)
}
\arguments{
\item{noalle}{noalle}
}
\description{
Lisää a priori jokaista alleelia joka populaation joka lokukseen
j 1/noalle(j) verran.
}
