% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIndLogml.R
\name{computeIndLogml}
\alias{computeIndLogml}
\title{computeIndLogml}
\usage{
computeIndLogml(omaFreqs, osuusTaulu)
}
\arguments{
\item{omaFreqs}{omaFreqs}

\item{osuusTaulu}{osuusTaulu}
}
\description{
Palauttaa yksilön logml:n, kun oletetaan yksilön alkuperät
määritellyiksi kuten osuusTaulu:ssa.
}
