% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePersonalAllFreqs.R
\name{computePersonalAllFreqs}
\alias{computePersonalAllFreqs}
\title{Compute Personal Freqs}
\usage{
computePersonalAllFreqs(ind, data, allFreqs, rowsFromInd, COUNTS = matrix(0))
}
\arguments{
\item{ind}{ind}

\item{data}{data}

\item{allFreqs}{allFreqs}

\item{rowsFromInd}{rowsFromInd}

\item{COUNTS}{COUNTS}
}
\description{
Laskee npops*(rowsFromInd*nloci) taulukon, jonka kutakin
saraketta vastaa yksilön ind alleeli. Eri rivit ovat alleelin
alkuperäfrekvenssit eri populaatioissa. Jos yksilölt?puuttuu jokin alleeli, 
niin vastaavaan kohtaa tulee sarake ykkösi?
}
