% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRows.R
\name{computeRows}
\alias{computeRows}
\title{Compute rows}
\usage{
computeRows(rowsFromInd, inds, ninds)
}
\arguments{
\item{rowsFromInd}{rowsFromInd}

\item{inds}{matrix}

\item{ninds}{ninds}
}
\description{
Individuals inds have been given. The function returns a vector,
containing the indices of the rows, which contain data from the individuals.
}
