% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_FASTA_to_BAPS.R
\name{convert_FASTA_to_BAPS}
\alias{convert_FASTA_to_BAPS}
\title{Convert an R object from FASTA to BAPS format}
\usage{
convert_FASTA_to_BAPS(data)
}
\arguments{
\item{data}{dataset to be converted}
}
\value{
`data` in BAPS format
}
\description{
Converts an R object from FASTA to BAPS format
}
\examples{
data <- system.file("extdata", "FASTA_clustering_diploid.fasta", package = "rBAPS")
convert_FASTA_to_BAPS(data)
}
\author{
Waldir Leoncio
}
