% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsiParas.R
\name{etsiParas}
\alias{etsiParas}
\title{Etsi Paras}
\usage{
etsiParas(osuus, osuusTaulu, omaFreqs, logml)
}
\arguments{
\item{osuus}{Percentages?}

\item{osuusTaulu}{Percentage table?}

\item{omaFreqs}{own Freqs?}

\item{logml}{log maximum likelihood}
}
\description{
Search for the best?
}
