% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgetl-fopen.R
\name{fgetl}
\alias{fgetl}
\title{Read line from file, removing newline characters}
\usage{
fgetl(file)
}
\arguments{
\item{file}{character vector to be read, usually an output of `fopen()`}
}
\value{
If the file is nonempty, then fgetl returns tline as a character vector. If the file is empty and contains only the end-of-file marker, then fgetl returns tline as a numeric value -1.
}
\description{
Equivalent function to its homonymous Matlab equivalent.
}
\seealso{
fopen
}
\author{
Waldir Leoncio
}
