% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find}
\alias{find}
\title{Find indices and values of nonzero elements}
\usage{
find(x, sort = TRUE)
}
\arguments{
\item{x}{object or logic operation on an object}

\item{sort}{sort output?}
}
\description{
Emulates behavior of `find`
}
