% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{fix}
\alias{fix}
\title{Round toward zero}
\usage{
fix(X)
}
\arguments{
\item{X}{input element}
}
\description{
Rounds each element of input to the nearest integer towards zero. Basically the same as trunc()
}
\author{
Waldir Leoncio
}
