% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgetl-fopen.R
\name{fopen}
\alias{fopen}
\title{Open file}
\usage{
fopen(filename)
}
\arguments{
\item{filename}{Path and name of file to be open}
}
\value{
The same as `readLines(filename)`
}
\description{
Open a text file
}
\seealso{
fgetl
}
\author{
Waldir Leoncio
}
