% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyPopMix.R
\name{greedyPopMix}
\alias{greedyPopMix}
\title{Clustering of pop individuals}
\usage{
greedyPopMix(data, format, partitionCompare = NULL, verbose = TRUE)
}
\arguments{
\item{data}{data file}

\item{format}{Data format. Format supported: "FASTA", "VCF" ,"BAM", "GenePop"}

\item{partitionCompare}{a properly-named list. Proper names include
"partitions"}

\item{verbose}{if \code{TRUE}, prints extra output information}
}
\description{
Clustering of pop individuals
}
\references{
Samtools: a suite of programs for interacting
with high-throughput sequencing data. <http://www.htslib.org/>
}
