% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleData.R
\name{handleData}
\alias{handleData}
\title{Handle Data}
\usage{
handleData(raw_data, format = "baps")
}
\arguments{
\item{raw_data}{Raw data in Genepop or BAPS format}

\item{format}{data format}
}
\description{
Handle Data
}
\details{
The last column of the original data tells you from which
individual that line is from. The function first examines how many line
maximum is from one individual giving know if it is haploid, diploid, etc.
After this function. Add blank lines for individuals with fewer rows as
maximum. If the code of an allele is = 0, the function changes that allele
code to the smallest code that is larger than any code in use. After this,
the function changes the allele codes so that one locus j
codes get values between? 1, ..., noalle(j).
}
