% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlePopData.R
\name{handlePopData}
\alias{handlePopData}
\title{Handle Pop data}
\usage{
handlePopData(raw_data)
}
\arguments{
\item{raw_data}{raw data}
}
\description{
The last column of the original data tells you which
<missing translation> that line is from. The function changes the allele
codes so that the codes for one locus have values between 1 and noalle[j].
Before this change, an allele whose code is zero is changed.
}
