% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFile.R
\name{importFile}
\alias{importFile}
\title{Import data file}
\usage{
importFile(data, format, verbose)
}
\arguments{
\item{data}{raw dataset}

\item{format}{data format (guesses from extension if not provided)}

\item{verbose}{if \code{TRUE}, prints extra output information}
}
\value{
The data in a format that can be used by the other functions
}
\description{
Imports data from several formats (FASTA, VCF, SAM, BAM).
}
\examples{
path_inst <- system.file("extdata", "", package = "rBAPS")
importFile(file.path(path_inst, "FASTA_clustering_haploid.fasta"))
}
