% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPopNames.R
\name{initPopNames}
\alias{initPopNames}
\title{Initialize Pop Names}
\usage{
initPopNames(nameFile, indexFile)
}
\arguments{
\item{nameFile}{nameFile}

\item{indexFile}{indexFile}
}
\description{
Initialize Pop Names
}
