% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputdlg.R
\name{inputdlg}
\alias{inputdlg}
\title{Gather user input}
\usage{
inputdlg(prompt, definput = NULL, dims = 1)
}
\arguments{
\item{prompt}{Text field with user instructions}

\item{definput}{default value of the input}

\item{dim}{number of dimensions in the answwers}
}
\description{
Replicates the functionality of the homonymous function in Matlab (sans dialog box)
}
