% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isfield.R
\name{isfield}
\alias{isfield}
\title{Checks if a list contains a field}
\usage{
isfield(x, field)
}
\arguments{
\item{x}{list}

\item{field}{name of field}
}
\description{
This function tries to replicate the behavior of the `isfield`
function in Matlab
}
\references{
https://se.mathworks.com/help/matlab/ref/isfield.html
}
