% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isGlobalEmpty.R
\name{isGlobalEmpty}
\alias{isGlobalEmpty}
\title{Check if global variable is empty}
\usage{
isGlobalEmpty(g)
}
\arguments{
\item{g}{the global variable in quesiton.}
}
\value{
TRUE if the variable still contains its original values, FALSE otherwise.
}
\description{
Checks if a global variable has been filled with values other than their initial ones.
}
\details{
For a list of global variables, check the \code{globals.R} file.
}
\author{
Waldir Leoncio
}
