% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isspace.R
\name{isspace}
\alias{isspace}
\title{Determine space characters}
\usage{
isspace(A)
}
\arguments{
\item{A}{a character array or a string scalar}
}
\value{
a vector TF such that the elements of TF are logical 1 (true) where corresponding characters in A are space characters, and logical 0 (false) elsewhere
}
\description{
Determine which characters are space characters
}
\note{
Recognized whitespace characters are ` ` and `\\t`.
}
\author{
Waldir Leoncio
}
