% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laskeOsaDist.R
\name{laskeOsaDist}
\alias{laskeOsaDist}
\title{Lower part of the dist}
\usage{
laskeOsaDist(inds2, dist, ninds)
}
\arguments{
\item{inds2}{inds2}

\item{dist}{dist}

\item{ninds}{ninds}
}
\description{
Constructs from the dist vector a subvector containing the individual inds2, Forms dist sub-vectors the vector, which includes yksiliden inds2
}
\author{
Waldir Leoncio
}
