% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix1-learn_partition_modified.R
\name{learn_partition_modified}
\alias{learn_partition_modified}
\title{Learn partition (modified)}
\usage{
learn_partition_modified(ordered)
}
\arguments{
\item{ordered}{ordered}
}
\value{
part
}
\description{
This function is called only if some individual has less than 
90 per cent non-missing data. The function uses fuzzy clustering for the
"non-missingness" values, finding maximum three clusters. If two of the
found clusters are such that all the values are >0.9, then those two are 
further combined.
}
