% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_simple_partition.R
\name{learn_simple_partition}
\alias{learn_simple_partition}
\title{Learn simple partition}
\usage{
learn_simple_partition(ordered_points, fii)
}
\arguments{
\item{ordered_points}{ordered_points}

\item{fii}{fii}
}
\description{
Goes through all the ways to divide the points into two or 
three groups. Chooses the partition which obtains highest logml.
}
