% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_gamma.R
\name{log_gamma}
\alias{log_gamma}
\title{Log Gamma}
\usage{
log_gamma(x)
}
\arguments{
\item{x}{number}
}
\value{
log(gamma(x)) for x > 0, Inf otherwise
}
\description{
Equal to log(gamma(x)) with special handling of x < 0 for
Matlab compatibility
}
