% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logml2String.R
\name{logml2String}
\alias{logml2String}
\title{Logml to string}
\usage{
logml2String(logml)
}
\arguments{
\item{logml}{input Logml}
}
\value{
String version of logml
}
\description{
Returns a string representation of a logml
}
