% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lueGenePopData.R
\name{lueGenePopData}
\alias{lueGenePopData}
\title{Read GenePop Data}
\usage{
lueGenePopData(tiedostonNimi)
}
\arguments{
\item{tiedostonNimi}{Name of the file}
}
\value{
list containing data and popnames
}
\description{
Reads GenePop-formatted data
}
