% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lueGenePopDataPop.R
\name{lueGenePopDataPop}
\alias{lueGenePopDataPop}
\title{Read GenePop Data}
\usage{
lueGenePopDataPop(tiedostonNimi)
}
\arguments{
\item{tiedostonNimi}{Name of the file}
}
\value{
List containing data and popnames
}
\description{
Read GenePop Data
}
\note{
The data is given in the form where the last column tells the
group. popnames are as before.
}
