% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lueNimi.R
\name{lueNimi}
\alias{lueNimi}
\title{Read the Name}
\usage{
lueNimi(line)
}
\arguments{
\item{line}{line}
}
\value{
nimi
}
\description{
Returns the part of the line from the beginning that is before the comma. Useful for returning the name of a GenePop area
}
