% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nargin.R
\name{nargin}
\alias{nargin}
\title{Number of function input arguments}
\usage{
nargin()
}
\value{
An integer
}
\description{
Returns the number of arguments passed to the parent function
}
\note{
This function only makes sense inside another function
}
\references{
https://stackoverflow.com/q/64422780/1169233
}
\author{
Waldir Leoncio
}
