% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noIndex.R
\name{noIndex}
\alias{noIndex}
\title{No index}
\usage{
noIndex(data, noalle)
}
\arguments{
\item{data}{data}

\item{noalle}{noalle}
}
\value{
puredata: a data contains no index column.
}
\description{
Checks that the data contains no index column.
}
\details{
As input, this function takes two variables from a mixture/admixture
result structure.
}
