% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownNum2Str.R
\name{ownNum2Str}
\alias{ownNum2Str}
\title{Own number to string}
\usage{
ownNum2Str(number)
}
\arguments{
\item{number}{number}
}
\description{
Converts numbers to strings
}
\note{
On Matlab, if number is NaN the output is 'NaN'. Here, the output will be an error. Also, the function belo expects "number" to have length one, whereas Matlab accepts vectors.
}
