% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion2str.R
\name{proportion2str}
\alias{proportion2str}
\title{Convert proportion to string}
\usage{
proportion2str(prob)
}
\arguments{
\item{prob}{belongs to [0.00, 0.01, ... ,1]}
}
\value{
a 4-mark presentation of proportion
}
\description{
Convert proportion to string
}
\note{
The `round` function in R, being ISO-compliant, rounds 8.5 to 8. The
Matlab equivalent rounds it to 9.
}
