% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Generate matrix with U(0, 1) trials}
\usage{
rand(r = 1, c = 1)
}
\arguments{
\item{r}{number of rows of output matrix}

\item{c}{number of columns of output matrix}
}
\value{
\eqn{r \times c} matrix with random trials from a standard uniform distribution.
}
\description{
Imitates the behavior of `rand()` on Matlab
}
