% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randga.R
\name{randga}
\alias{randga}
\title{Generates random number from a Gamma distribution}
\usage{
randga(a, b)
}
\arguments{
\item{a}{shape}

\item{b}{rate}
}
\value{
One realization of Gamma(a, b)
}
\description{
Generates one random number from shape parameter a and rate
parameter b
}
\details{
The generated random variable has mean a / b. It will be
positively-skewed for small values, but converges to a symmetric distribution
for very large numbers of a and b.
}
