% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape}
\alias{reshape}
\title{Reshape array}
\usage{
reshape(A, sz)
}
\arguments{
\item{A}{input matrix}

\item{sz}{vector containing the dimensions of the output vector}
}
\description{
Reshapes a matrix according to a certain number of dimensions
}
\details{
This function replicates the functionality of the `reshape()`
function on Matlab. This function is basically a fancy wrapper for the
`array()` function in R, but is useful because it saves the user translation
time. Moreover, it introduces validation code that alter the behavior of
`array()` and makes it more similar to `replicate()`.
}
\note{
The Matlab function also accepts as input the dismemberment of sz as
scalars.
}
