% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selvitaDigitFormat.R
\name{selvitaDigitFormat}
\alias{selvitaDigitFormat}
\title{Find out the Digit Format}
\usage{
selvitaDigitFormat(line)
}
\arguments{
\item{line}{the first line after the "pop" word from data in Genepop format. #  @note Function clarified based on the line format whether the alleles of the data are given using 2 or 3 numbers.}
}
\value{

}
\description{
Find out the Digit Format
}
