% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\name{setdiff_MATLAB}
\alias{setdiff_MATLAB}
\title{Set differences of two arrays}
\usage{
setdiff_MATLAB(A, B, legacy = FALSE)
}
\arguments{
\item{A}{first array}

\item{B}{second awway}

\item{legacy}{if `TRUE`, preserves the behavior of}
}
\value{

}
\description{
Loosely replicates the behavior of the homonym Matlab function
}
\author{
Waldir Leoncio
}
