% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateAllFreqs.R
\name{simulateAllFreqs}
\alias{simulateAllFreqs}
\title{Simulate All Frequencies}
\usage{
simulateAllFreqs(noalle, COUNTS = matrix(0))
}
\arguments{
\item{noalle}{noalle}

\item{COUNTS}{COUNTS}
}
\description{
Lisää jokaista alleelia joka populaation joka lokukseen j1/noalle(j) verran. Näin saatuja counts:eja vastaavista Dirichlet-jakaumista simuloidaan arvot populaatioiden alleelifrekvensseille.
Add each allele to each locus in each population by j 1 / noalle(j). The Dirichlet distributions corresponding to the counts thus obtained simulate values for the allele frequencies of the populations.
}
