% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuloiAlleeli.R
\name{simuloiAlleeli}
\alias{simuloiAlleeli}
\title{simuloiAlleeli}
\usage{
simuloiAlleeli(allfreqs, pop, loc)
}
\arguments{
\item{allfreqs}{allfreqa}

\item{pop}{pop}

\item{loc}{loc}
}
\description{
Simuloi populaation pop lokukseen loc alleelin.
}
\note{
This function is (only?) called by `simulateIndividuals()`. Therefore,
exporting it is probably unnecessary.
}
