% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortrows.R
\name{sortrows}
\alias{sortrows}
\title{Sort rows of matrix or table}
\usage{
sortrows(A, column = 1)
}
\arguments{
\item{A}{matrix}

\item{column}{ordering column}
}
\description{
Emulates the behavior of the `sortrows` function on Matlab
}
