% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strcmp.R
\name{strcmp}
\alias{strcmp}
\title{Compare two character elements}
\usage{
strcmp(s1, s2)
}
\arguments{
\item{s1}{first character element (string, vector or matrix)}

\item{s2}{second character element (string, vector or matrix)}
}
\value{
a logical element of the same type as the input
}
\description{
Logical test if two character elements are identical
}
