% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeLine.R
\name{takeLine}
\alias{takeLine}
\title{Take line}
\usage{
takeLine(description, width)
}
\arguments{
\item{description}{description}

\item{width}{width}
}
\value{
newline
}
\description{
Returns one line from the description.
}
