% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testaaGenePopData.R
\name{testaaGenePopData}
\alias{testaaGenePopData}
\title{Tests GenePop data}
\usage{
testaaGenePopData(tiedostonNimi)
}
\arguments{
\item{tiedostonNimi}{Filename}
}
\value{
kunnossa (binary "ok" condition value) == 0 if the data is not valid genePop data. Otherwise, kunnossa == 1.
}
\description{
Tests GenePop data
}
\details{
GenePop data are textfiles that follow the GenePop format. This function checks if such file is properly formatted as GenePop.
}
