% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testaaKoordinaatit.R
\name{testaaKoordinaatit}
\alias{testaaKoordinaatit}
\title{Test the coordinates}
\usage{
testaaKoordinaatit(ninds, coordinates)
}
\arguments{
\item{ninds}{ninds}

\item{coordinates}{coordinates}
}
\value{
a list of defectives ("viallinen") and coordinates
}
\description{
Test the coordinates
}
