% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testaaOnkoKunnollinenBapsData.R
\name{testaaOnkoKunnollinenBapsData}
\alias{testaaOnkoKunnollinenBapsData}
\title{TestBAPS data}
\usage{
testaaOnkoKunnollinenBapsData(data)
}
\arguments{
\item{data}{dataset}
}
\value{
ninds
}
\description{
Test if loaded BAPS data is proper
}
