% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testaaPop.R
\name{testaaPop}
\alias{testaaPop}
\title{Test population}
\usage{
testaaPop(rivi)
}
\arguments{
\item{rivi}{Line}
}
\value{
pal = 1 if the line starts with one of the following
}
\description{
Test a line in the population
}
