% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{times}
\alias{times}
\title{Element-wise matrix multiplication}
\usage{
times(a, b)
}
\arguments{
\item{a}{first factor of the multiplication}

\item{b}{second factor of the multiplication}
}
\value{
matrix with dimensions equal to the larger of the two factors
}
\description{
Emulates the `times()` and `.*` operators from Matlab.
}
\details{
This function basically handles elements of different length better than the `*` operator in R, at least as far as behavior from a Matlab user is expecting.
}
