% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiputfile.R
\name{uiputfile}
\alias{uiputfile}
\title{Save file}
\usage{
uiputfile(filter = ".rda", title = "Save file")
}
\arguments{
\item{filter}{accepted file extension}

\item{title}{Title}
}
\description{
This function intends to loosely mimic the behaviour of the
homonymous Matlab function.
}
