% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMixtureInfoPop.R
\name{writeMixtureInfoPop}
\alias{writeMixtureInfoPop}
\title{Write Mixture Info Pop}
\usage{
writeMixtureInfoPop(
  logml,
  rows,
  data,
  adjprior,
  priorTerm,
  outPutFile,
  inputFile,
  partitionSummary,
  popnames,
  fixedK
)
}
\arguments{
\item{logml}{logml}

\item{rows}{rows}

\item{data}{data}

\item{adjprior}{adjprior}

\item{priorTerm}{priorTerm}

\item{outPutFile}{outPutFile}

\item{inputFile}{inputFile}

\item{partitionSummary}{partitionSummary}

\item{popnames}{popnames}

\item{fixedK}{fixedK}
}
\description{
Writes information about the pop mixture
}
