% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros_ones.R
\name{zeros}
\alias{zeros}
\title{Matrix of zeros}
\usage{
zeros(n1, n2 = n1)
}
\arguments{
\item{n1}{number of rows}

\item{n2}{number of columns}
}
\description{
wrapper of `zeros_or_ones()` that replicates the behavior of 
the `zeros()` function on Matlab
}
