% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix1.R
\name{admix1}
\alias{admix1}
\title{Admixture analysis}
\usage{
admix1(tietue)
}
\arguments{
\item{tietue}{tietue}
}
\description{
Admixture analysis
}
\details{
If the record == -1, the mixture results file is loaded. Otherwise,
will the required variables be retrieved from the record fields?
`tietue`should contain the following elements: PARTITION, COUNTS, SUMCOUNTS,
alleleCodes, adjprior, popnames, rowsFromInd, data, npops, noalle
}
