% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_FASTA_to_BAPS.R
\name{convert_FASTA_to_BAPS}
\alias{convert_FASTA_to_BAPS}
\title{Convert from FASTA to BAPS}
\usage{
convert_FASTA_to_BAPS(file)
}
\arguments{
\item{file}{filename of FASTA file}
}
\value{
`data` in BAPS format
}
\description{
Converts a file (not an R object) from FASTA to BAPS format
}
\examples{
file <- system.file("extdata", "FASTA_clustering_haploid.fasta", package = "rBAPS")
convert_FASTA_to_BAPS(file)
}
\author{
Waldir Leoncio
}
