% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyMix.R
\name{greedyMix}
\alias{greedyMix}
\title{Clustering of individuals}
\usage{
greedyMix(data, format)
}
\arguments{
\item{data}{data file}

\item{format}{Format of the data c("FASTA", "VCF" ,"SAM", or "GenePop")}
}
\description{
Clustering of individuals
}
