% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyMix.R
\name{greedyMix}
\alias{greedyMix}
\title{Clustering of individuals}
\usage{
greedyMix(
  data,
  format,
  c.rows,
  partitionCompare.partitions,
  ninds,
  rowsFromInd,
  noalle,
  adjprior,
  priorTerm,
  alleleCodesinp,
  popnames,
  fixedK = FALSE,
  partition_compare = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{data file}

\item{format}{Data format. Format supported: "FASTA", "VCF" ,"BAM", "GenePop"}

\item{verbose}{if \code{TRUE}, prints extra output information}
}
\description{
Clustering of individuals
}
\examples{
data <- system.file("extdata", "FASTA_clustering_haploid.fasta", package = "rBAPS")
greedyMix(data)
}
\references{
Samtools: a suite of programs for interacting
with high-throughput sequencing data. <http://www.htslib.org/>
}
