% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyMix.R
\name{greedyMix}
\alias{greedyMix}
\title{Clustering of individuals}
\usage{
greedyMix(
  tietue,
  format = NULL,
  savePreProcessed = NULL,
  filePreProcessed = NULL
)
}
\arguments{
\item{tietue}{File}

\item{format}{Format of the data ("BAPS", "GenePop" or "Preprocessed")}

\item{savePreProcessed}{Save the pre-processed data?}

\item{filePreProcessed}{Is the file already processed?}
}
\description{
Clustering of individuals
}
