% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyMix.R
\name{greedyMix}
\alias{greedyMix}
\title{Clustering of individuals}
\usage{
greedyMix(data, format, verbose = TRUE)
}
\arguments{
\item{data}{data file}

\item{format}{Format of the data c("FASTA", "VCF" ,"BAM", or "GenePop")}

\item{verbose}{if \code{TRUE}, prints extra output information}
}
\description{
Clustering of individuals
}
\references{
Samtools: a suite of programs for interacting
with high-throughput sequencing data. <http://www.htslib.org/>
}
