% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage}
\alias{linkage}
\title{Linkage}
\usage{
linkage(Y, method = "co")
}
\arguments{
\item{Y}{matrix}

\item{method}{either 'si', 'av', 'co' 'ce' or 'wa'}
}
\description{
Create hierarchical cluster tree.
}
\details{
Z = LINKAGE(Y) creates a hierarchical cluster tree, using the single
linkage algorithm.  The input Y is a distance matrix such as is generated by
PDIST. Y may also be a more general dissimilarity matrix conforming to the
output format of PDIST.

Z = linkage(X) returns a matrix Z that encodes a tree containing hierarchical clusters of the rows of the input data matrix X.
}
\note{
This is also a base MATLAB function. The reason why the BAPS
source code also contains a LINKAGE function is unclear. One could speculate
that BAPS should use this function instead of the base one, so this is why
this function is part of this package (instead of a MATLAB-replicating
package such as matlab2r)
}
