% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab2r.R
\name{matlab2r}
\alias{matlab2r}
\title{Convert Matlab function to R}
\usage{
matlab2r(filename, saveOutput = FALSE)
}
\arguments{
\item{filename}{name of the file}

\item{saveOutput}{if `TRUE`, `filename` is overwritten. Defaults to `FALSE`}
}
\value{
text converted to R, printed to screen or replacing input file
}
\description{
Performs basic syntax conversion from Matlab to R
}
\author{
Waldir Leoncio
}
