% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab2r.R
\name{matlab2r}
\alias{matlab2r}
\title{Convert Matlab function to R}
\usage{
matlab2r(filename, output = "clean", improve_formatting = TRUE)
}
\arguments{
\item{filename}{name of the file}

\item{output}{can be "asis", "clean" (default) or "save"}

\item{improve_formatting}{if `TRUE` (default), makes minor changes
to conform to best-practice formatting conventions}
}
\value{
text converted to R, printed to screen or replacing input file
}
\description{
Performs basic syntax conversion from Matlab to R
}
\author{
Waldir Leoncio
}
