% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poistaLiianPienet.R
\name{poistaLiianPienet}
\alias{poistaLiianPienet}
\title{Remove too small}
\usage{
poistaLiianPienet(
  npops,
  rowsFromInd,
  alaraja,
  PARTITION = matrix(NA, 0, 0),
  COUNTS = matrix(NA, 0, 0),
  SUMCOUNTS = NA
)
}
\arguments{
\item{npops}{npops}

\item{rowsFromInd}{rowsFromInd}

\item{alaraja}{alaraja}
}
\description{
Muokkaa tulokset muotoon, jossa outlier yksilöt on poistettu.
Tarkalleen ottaen poistaa ne populaatiot, joissa on vähemmän kuin
'alaraja':n verran yksilöit?
}
