% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randdir.R
\name{randdir}
\alias{randdir}
\title{Generates random numbers}
\usage{
randdir(counts, nc)
}
\arguments{
\item{counts}{shape parameter}

\item{nc}{number of rows on output}
}
\value{
vector of length `nc` with r.v. realizations from Gamma(rate=1)
}
\description{
Generates random numbers
}
\examples{
randdir(matrix(c(10, 30, 60), 3), 3)
}
\seealso{
randga
}
