% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBAPS-package.R
\docType{package}
\name{rBAPS}
\alias{rBAPS}
\title{Bayesian Analysis of Population Structure}
\description{
This is a partial implementation of the BAPS software
}
\note{
Found a bug? Want to suggest a feature? Contribute to the scientific
and open source communities by opening an issue on our home page.
Check the "BugReports" field on the package description for the URL.
}
