% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivinSisaltamienMjonojenLkm.R
\name{rivinSisaltamienMjonojenLkm}
\alias{rivinSisaltamienMjonojenLkm}
\title{Number of M queues}
\usage{
rivinSisaltamienMjonojenLkm(line)
}
\arguments{
\item{line}{line number}
}
\value{
count
}
\description{
Number of M queues
}
