% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIndividuals.R
\name{simulateIndividuals}
\alias{simulateIndividuals}
\title{Simulate individuals}
\usage{
simulateIndividuals(n, rowsFromInd, allfreqs, pop, missing_level)
}
\arguments{
\item{n}{n}

\item{rowsFromInd}{rowsFromInd}

\item{allfreqs}{allfreqs}

\item{pop}{pop}

\item{missing_level}{missing_level}
}
\description{
simulate n individuals from population pop, such that
proportion "missing_level" of the alleles are present.
}
