% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laskeMuutokset12345.R
\name{spatialMixture_muutokset}
\alias{spatialMixture_muutokset}
\title{Calculate changes (spatial mixture class)}
\description{
Calculate changes (spatial mixture class)

Calculate changes (spatial mixture class)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spatialMixture_muutokset-laskeMuutokset}{\code{spatialMixture_muutokset$laskeMuutokset()}}
\item \href{#method-spatialMixture_muutokset-clone}{\code{spatialMixture_muutokset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialMixture_muutokset-laskeMuutokset"></a>}}
\if{latex}{\out{\hypertarget{method-spatialMixture_muutokset-laskeMuutokset}{}}}
\subsection{Method \code{laskeMuutokset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialMixture_muutokset$laskeMuutokset(
  ind,
  rowsFromInd,
  data,
  adjprior,
  priorTerm,
  logml,
  cliques,
  separators
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind}}{ind}

\item{\code{rowsFromInd}}{rowsFromInd}

\item{\code{data}}{data}

\item{\code{adjprior}}{adjprior}

\item{\code{priorTerm}}{priorTerm}

\item{\code{logml}}{logml}

\item{\code{cliques}}{cliques}

\item{\code{separators}}{separators}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialMixture_muutokset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spatialMixture_muutokset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialMixture_muutokset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
