% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suoritaMuutos.R
\name{suoritaMuutos}
\alias{suoritaMuutos}
\title{suoritaMuutos}
\usage{
suoritaMuutos(osuusTaulu, osuus, indeksi, COUNTS = matrix(0))
}
\arguments{
\item{osuusTaulu}{Percentage table?}

\item{osuus}{percentage?}

\item{indeksi}{index}

\item{COUNTS}{counts}
}
\description{
P채ivitt채채 osuusTaulun muutoksen j채lkeen.
}
