% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMixtureInfo.R
\name{writeMixtureInfo}
\alias{writeMixtureInfo}
\title{Write Mixture Info}
\usage{
writeMixtureInfo(
  logml,
  rowsFromInd,
  data,
  adjprior,
  priorTerm,
  outPutFile,
  inputFile,
  partitionSummary,
  popnames,
  fixedK,
  verbose
)
}
\arguments{
\item{logml}{logml}

\item{rowsFromInd}{rowsFromInd}

\item{data}{data}

\item{adjprior}{adjprior}

\item{priorTerm}{priorTerm}

\item{outPutFile}{outPutFile}

\item{inputFile}{inputFile}

\item{partitionSummary}{partitionSummary}

\item{popnames}{popnames}

\item{fixedK}{fixedK}

\item{verbose}{if \code{TRUE}, prints extra output information}
}
\description{
Writes information about the mixture
}
